third_party/WebKit/LayoutTests/inspector/console/console-tests-expected.txt

CONSOLE MESSAGE: line 9: log
CONSOLE DEBUG: line 10: debug
CONSOLE INFO: line 11: info
CONSOLE WARNING: line 12: warn
CONSOLE ERROR: line 13: error
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 15: repeated
CONSOLE MESSAGE: line 20: 1
CONSOLE MESSAGE: line 22: [object Object]
CONSOLE MESSAGE: line 23: 42
CONSOLE MESSAGE: line 24: true
CONSOLE MESSAGE: line 25: null
CONSOLE MESSAGE: line 26: undefined
CONSOLE MESSAGE: line 27: [object HTMLDocument]
CONSOLE MESSAGE: line 28: function () { }
CONSOLE MESSAGE: line 29: function f() { }
CONSOLE MESSAGE: line 30: 1,2,3
CONSOLE MESSAGE: line 31: /regexp.*/
Tests that console logging dumps proper messages.

log console-tests.html:9 console-message console-log-level
debug console-tests.html:10 console-message console-debug-level
info console-tests.html:11 console-message console-info-level
warn console-tests.html:12 console-message console-warning-level
error console-tests.html:13 console-message console-error-level
5repeated console-tests.html:15 console-message console-log-level repeated-message
count: 1 console-tests.html:17 console-message console-debug-level
count: 2 console-tests.html:17 console-message console-debug-level
group console-tests.html:18 console-message console-log-level console-group-title
1 2 3 console-tests.html:20 console-message console-log-level
groupCollapsed console-tests.html:21 console-message console-log-level console-group-title
: 1 console-tests.html:33 console-message console-debug-level
: 1 console-tests.html:34 console-message console-debug-level
: 1 console-tests.html:35 console-message console-debug-level
title: 1 console-tests.html:36 console-message console-debug-level
title: 2 console-tests.html:37 console-message console-debug-level
title: 3 console-tests.html:38 console-message console-debug-level
